[FIX] l10n_es_edi_tbai: omit recipient for simplified invoices
Since Odoo 18.0, the recipient is always included when sending
a simplified invoice. While this is allowed by TicketBAI, it is
only valid if the partner has a VAT number (NIF).

Steps to reproduce:

Install the l10n_es_edi_tbai module and setup an ES company
Create a simplified invoice with the partner "Simplified Invoice Partner (ES)"
Send the invoice to TicketBAI
The following error will be returned:
"B4_1000002: Todos los registros incluidos en la petición son
incorrectos."

The Fix:
For simplified invoices, this fix ensures that the recipient is only
included in the XML if the partner has a VAT/NIF. If not, the recipient
is omitted to avoid TicketBAI validation errors.

opw-4525875
opw-4522938
opw-4553030

closes odoo/odoo#198156

X-original-commit: 2d68a08b4a2ace59b3598f8d3d676be6da9703e0
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: Louis Gobert (loug) <loug@odoo.com>